Remove theme config parts. One object with all configuration is easier to work with since theme config is not massive
