--- src/theme/parts/breakpoints.ts
@@ -1,5 +1,0 @@
-import { Breakpoints } from '../../types/theme/Breakpoints';
-
-export type BreakpointScaleIndex = 0 | 1 | 2
-
-export const breakpoints: Breakpoints = [640, 768, 1024];
