--- src/theme/parts/layout.ts
@@ -1,3 +1,0 @@
-export const layout = {
-  siteWidth: '800px',
-};
