--- src/theme/parts/spacings.ts
@@ -1,5 +1,0 @@
-import { SpacingScale } from '../../types/theme/Spacing';
-
-const spacingScale: SpacingScale = [5, 10, 20, 40];
-
-export const spacings = spacingScale;
