+++ src/types/theme/Breakpoints.ts
@@ -1,1 +1,3 @@
 export type Breakpoints = number[]
+
+export type BreakpointScaleIndex = 0 | 1 | 2
